adding author id verification

+++ src/modules/news/services/DeleteNewsService.ts
@@ -10,11 +10,18 @@ class DeleteNewsService {
     private newsRepository: INewsRepository,
   ) {}
 
-  public async execute(id: string): Promise<void> {
+  public async execute(id: string, authorId: string): Promise<void> {
     const news = await this.newsRepository.findById(id);
 
     if (!news) {
       throw new AppError('Notícia não encontrada', 404);
+    }
+
+    if (news.authorId !== authorId) {
+      throw new AppError(
+        'Você não tem autorização para editar essa notícia',
+        403,
+      );
     }
 
     await this.newsRepository.delete(id);
